+++ HISTORY.md
@@ -1,3 +1,8 @@
+## v5.1.1-beta-1
+* New RuleFor(x => x.prop, 'constant')
+* Support for Hashids.net: RuleFor(x => x.Id, f => f.Hashids.Encode())
+* 
+
 ## v5.0.1
 * Roll-up Release for .NET Framework since v4.0.1.
 * CoreCLR users please continue using latest beta release until CoreCLR is RTM.